Factor out the conditional extensions from legacy.
