+++ lib/extensions/conditional.rb
@@ -1,0 +1,1 @@
+module Extensions::Conditional; end
