+++ lib/extensions/conditional/active_record.rb
@@ -1,0 +1,1 @@
+module Extensions::Conditional::ActiveRecord; end
