+++ lib/extensions/legacy/active_record/base.rb
@@ -15,25 +15,5 @@ module Extensions::Legacy::ActiveRecord::Base
     def acts_as_lesson_plan_item
       acts_as :lesson_plan_item, class_name: Course::LessonPlanItem.name
     end
-
-    # Functions from conditional-and-condition framework.
-    # Declare this function in the conditional model that requires conditions.
-    def acts_as_conditional
-      has_many :conditions, -> { includes :actable },
-               class_name: Course::Condition.name, as: :conditional, dependent: :destroy
-
-      include ConditionalInstanceMethods
-    end
-
-    # Declare this function in the condition model
-    def acts_as_condition
-      acts_as :condition, class_name: Course::Condition.name
-    end
-  end
-
-  module ConditionalInstanceMethods
-    def specific_conditions
-      conditions.map(&:actable)
-    end
   end
 end
